requirements.txt
google-cloud
google-cloud-vision
pyzmq
google-generativeai
torch
torchvision
torchaudio
transformers
python-dotenv
streamlit
